Remove the backup files & updates to main file
